requirements.txt
google-genai>=1.52.0
streamlit>=1.28.0
requests>=2.31.0
numpy>=1.24.0
sounddevice>=0.4.6
python-dotenv>=1.0.0
